- changed the hook so that it handles fetching and filtering; - changed MovieInfo component, that it is solely responsible for rendering and displaying
